Solution for day 13
Fun to run a board game on the int computer. Made mine so that I can
show the game while the program is playing. I did have a manual mode
at first, but it would have taken forever to play and the program did
not allow any mistakes at times.