Solution for day 25
Semi-manual solution. Fun to have the last puzzle of the year be
a text based game. Tried to solve it all manually at first, but
did not have the patience to solve it all myself, so added a
method on my droid to try all the solutions after I manually find
and set up what was needed before that.